caa_web_api/requirements.txt
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.2
django-crispy-forms==1.7.2
djangorestframework==3.8.2
httpie==0.9.9
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
Pygments==2.2.0
pytz==2018.5
requests==2.19.1
uritemplate==3.0.0
urllib3==1.23
